Floating point numbers should be allowed for step

+++ src/Fields/Attributes/Step.php
@@ -15,7 +15,7 @@ namespace WordPlate\Acf\Fields\Attributes;
 
 trait Step
 {
-    public function step(int $step): self
+    public function step(float $step): self
     {
         $this->config->set('step', $step);
